vim: fix asyncomplete when cpu load is high
Especially when typing <BS><BS> quickly, the completion would insert the
current word. The issue is that setlocal completeopt would be set too
fast so it would contain insert and select. This would result in
inserting the first match.
Now after 200ms the auto popup is turned off (if pum is still not
visible) and after another 1000ms the completeopt is reset. As it is
unrealistic to complete and trigger the next completion within 1.2
seconds this should be fine. Otherwise the delay could be reduced.