git: hook ignore wip/fixup after 14 days
When a colleague (or myself) pushed a wip or fixup commit which can't be
cleaned anymore it would result in not being able to push any new branch
ever again.
Old branches would not matter as there only a range is checked but for
new branches (not pushed to the remote) the whole history is checked.
Now only the commits of the last 14 days are checked when a new branch
is pushed. This should include the whole pushed branch.