vim: define Leader format and code as function
This allows to overwrite the function in an ftplugin. Before the
mapping would have been overwritten. Resulting in searching every file
if contains a leader mapping.
Maybe some mappings like compile, test and run should be a LocalLeader
instead? Format could stay as Leader mapping as it is general purpose
and only overwritten if there is a specific tool for this language.