Drop vim support - only neovim from now
While this does not change much right now, it means that I won't try to
limit neovim in order to have vim working as well. Since I now use VimR
as gui and solely neovim on the terminal there is no reason to support
both.

To be 100% honest, I haven't really tested Vim anyways. So this is more
of a change of mindset. I'm not sure how to proceed with ~/.vim. Maybe I
rename it to ~/.nvim. I'm not really fond of ~/.config/nvim but maybe
this will change?