vim: add Amake command and improve qf open
Also add a wrapper around asyncdo to be able to run stop and run with a
given type (l|c). This makes it easier to share code for things like
LGrep and Grep or Lmake and Amake.

Also opening the qf list now checks for the right type of qf list. So
Lmake won't open the location list when Grep is done filling the
quickfix list.