vim: replace `fugitive#head` with `gitsigns.head`
In the future I want to parse the status on my own in order to show more
granular information like detached or the current tag. (Which is neither
possible with gitsigns nor with fugitive - gitsigns is closer but still
not there.)

Also lazy load fugitive now since I don't require it for my statusline
anymore.